# sisyphus-server/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sisyphus Server API",
    "description": "自托管 CI 平台 Server REST API。业务端点均在 /api/v1/ 前缀下；B2b 起会话认证全局把关（未认证一律 401，放行面仅 login/setup/healthz 与静态资源，票 B2b-T1）。",
    "contact": {
      "name": "sisyphus contributors"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{name}/pipelines/{pipeline}/builds": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "构建列表（viewer 档）：按号倒序 + 分页 + 状态过滤。",
        "operationId": "list",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "项目名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline",
            "in": "path",
            "description": "pipeline 名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "单页条数（1..=100，默认 20）。",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码（从 1 起，默认 1）。",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "running",
                "succeeded",
                "failed",
                "cancelled",
                "timeout"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "构建概要列表（按号倒序）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "权限不足（列表需 viewer 档）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "项目不存在/不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "分页/状态过滤参数非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "builds"
        ],
        "summary": "手动触发构建（runner 档）：参数覆盖默认值 + 可选分支/commit/revision，\n调 engine 统一触发入口，返回 202 + 构建号。",
        "operationId": "trigger",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "项目名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline",
            "in": "path",
            "description": "pipeline 名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "已触发，返回构建号（异步推进）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildAcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "viewer 档不足（触发需 runner 档）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "项目不存在/不可见，或 pipeline 不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "请求体非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildAcceptedResponse": {
        "type": "object",
        "description": "触发 / 重跑的受理响应：构建号 + 当前状态（异步推进，sched 驱动）。",
        "required": [
          "number",
          "build_id",
          "attempt",
          "status"
        ],
        "properties": {
          "attempt": {
            "type": "integer",
            "format": "int32",
            "description": "重跑 attempt（首跑 1；from_failed +1）。"
          },
          "build_id": {
            "type": "integer",
            "format": "int64",
            "description": "构建行 id。"
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "description": "per-pipeline 构建号。"
          },
          "status": {
            "$ref": "#/components/schemas/BuildStatusDto",
            "description": "受理时状态（queued / running）。"
          }
        }
      },
...
      "BuildListResponse": {
        "type": "object",
        "description": "分页列表响应。",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuildSummaryResponse"
            },
            "description": "本页构建概要（按号倒序）。"
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "单页条数。"
          },
          "page": {
            "type": "integer",
            "format": "int64",
            "description": "当前页码。"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "总条数（满足过滤条件；供客户端算页数）。"
          }
        }
      },
      "BuildStatusDto": {
        "type": "string",
        "description": "构建状态（REST 形态；落库文本同型）。",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled",
          "timeout"
        ]
      },
      "BuildSummaryResponse": {
        "type": "object",
        "description": "列表条目（构建概要）。",
        "required": [
          "number",
          "pipeline_name",
          "status",
          "trigger",
          "trigger_by",
          "attempt"
        ],
        "properties": {
          "attempt": {
            "type": "integer",
            "format": "int32",
            "description": "重跑 attempt。"
          },
          "cancelled_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "取消时刻。"
          },
          "finished_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "终态时刻。"
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "description": "per-pipeline 构建号。"
          },
          "pipeline_name": {
            "type": "string",
            "description": "pipeline 名。"
          },
          "started_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "开始时刻（queued→running；未运行为空）。"
          },
          "status": {
            "$ref": "#/components/schemas/BuildStatusDto",
            "description": "构建状态。"
          },
          "trigger": {
            "$ref": "#/components/schemas/TriggerSourceDto",
            "description": "触发源。"
          },
          "trigger_by": {
            "type": "string",
            "description": "触发人（业务表实名）。"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "统一错误响应体。",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码（大写蛇形，如 `NOT_FOUND`、`VALIDATION_FAILED`）。"
          },
          "detail": {
            "description": "结构化补充（如校验错误清单）；无补充时缺省不输出。"
          },
          "message": {
            "type": "string",
            "description": "人读错误信息，可直接展示。"
          }
        }
      },
...
      "TriggerBuildRequest": {
        "type": "object",
        "description": "手动触发请求体：参数覆盖（默认值之上叠加）+ 可选 git 分支/commit 或 svn\nrevision。空体等价于全默认（参数取定义默认值、git 缺省项目默认分支）。",
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "git 分支（手动可选；缺省项目默认分支；svn 无分支概念）。"
          },
          "commit": {
            "type": [
              "string",
              "null"
            ],
            "description": "git commit sha（手动未钉为空——Agent 检分支头）。"
          },
          "params": {
            "type": "object",
            "description": "参数覆盖（名→值；默认值之上叠加，不存在的名无害——未被引用即无效果）。",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "revision": {
            "type": [
              "string",
              "null"
            ],
            "description": "svn revision（手动可选；git 为空）。"
          }
        }
      },
...
      "TriggerSourceDto": {
        "type": "string",
        "description": "构建触发源（手动/cron/poll）。",
        "enum": [
          "manual",
          "cron",
          "poll"
        ]
      },
...
    }
  },
...
}